Move tests to sub-directory: o1-utils
