+++ utils/src/math.rs
@@ -21,26 +21,3 @@ pub fn ceil_log2(d: usize) -> usize {
 pub fn div_ceil(a: usize, b: usize) -> usize {
     (a + b - 1) / b
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_log2() {
-        let tests = [
-            (1, 0),
-            (2, 1),
-            (3, 2),
-            (9, 4),
-            (15, 4),
-            (15430, 14),
-            (usize::MAX, 64),
-        ];
-        for (d, expected_res) in tests.iter() {
-            let res = ceil_log2(*d);
-            println!("ceil(log2({d})) = {res}, expected = {expected_res}");
-            assert!(res == *expected_res)
-        }
-    }
-}
